Refactor ProposalActions components
